refactor the constants into separate files
