+++ src/device-registry/config/environments/development.js
@@ -1,0 +1,28 @@
+const mongoose = require("mongoose");
+const ObjectId = mongoose.Types.ObjectId;
+const log4js = require("log4js");
+const isEmpty = require("is-empty");
+const { logText } = require("@utils/log");
+const logger = log4js.getLogger(`${this.ENVIRONMENT} -- constants-config`);
+
+const devConfig = {
+  DEFAULT_COHORT: process.env.DEV_DEFAULT_COHORT,
+  MONGO_URI: process.env.MONGO_URI_DEV,
+  DB_NAME: process.env.MONGO_DEV,
+  REDIS_SERVER: process.env.REDIS_SERVER_DEV,
+  REDIS_PORT: process.env.REDIS_PORT,
+  KAFKA_BOOTSTRAP_SERVERS: process.env.KAFKA_BOOTSTRAP_SERVERS_DEV
+    ? process.env.KAFKA_BOOTSTRAP_SERVERS_DEV.split(",").filter(
+        (value) => value.trim() !== ""
+      )
+    : [],
+  KAFKA_TOPICS: process.env.KAFKA_TOPICS_DEV,
+  SCHEMA_REGISTRY: process.env.SCHEMA_REGISTRY_DEV,
+  KAFKA_RAW_MEASUREMENTS_TOPICS: process.env.KAFKA_RAW_MEASUREMENTS_TOPICS_DEV,
+  KAFKA_CLIENT_ID: process.env.KAFKA_CLIENT_ID_DEV,
+  KAFKA_CLIENT_GROUP: process.env.KAFKA_CLIENT_GROUP_DEV,
+  ENVIRONMENT: "DEVELOPMENT ENVIRONMENT",
+  DATAWAREHOUSE_METADATA: process.env.DATAWAREHOUSE_METADATA_DEV,
+  DATAWAREHOUSE_AVERAGED_DATA: process.env.DATAWAREHOUSE_AVERAGED_DATA_DEV,
+};
+module.exports = devConfig;
