+++ src/device-registry/config/global/numericals.js
@@ -1,0 +1,15 @@
+const mongoose = require("mongoose");
+const ObjectId = mongoose.Types.ObjectId;
+const log4js = require("log4js");
+const isEmpty = require("is-empty");
+const { logText } = require("@utils/log");
+const logger = log4js.getLogger(`${this.ENVIRONMENT} -- constants-config`);
+
+const numericals = {
+  MQTT_BRIDGE_PORT: 8883,
+  NUM_MESSAGES: 5,
+  TOKEN_EXP_MINS: 360,
+  MINIMUM_BACKOFF_TIME: 1,
+  MAXIMUM_BACKOFF_TIME: 32,
+};
+module.exports = numericals;
