+++ src/device-registry/config/global/query-limits.js
@@ -1,0 +1,30 @@
+const mongoose = require("mongoose");
+const ObjectId = mongoose.Types.ObjectId;
+const log4js = require("log4js");
+const isEmpty = require("is-empty");
+const { logText } = require("@utils/log");
+const logger = log4js.getLogger(`${this.ENVIRONMENT} -- constants-config`);
+
+const queryLimits = {
+  DEFAULT_LIMIT_FOR_QUERYING_SITES:
+    process.env.DEFAULT_LIMIT_FOR_QUERYING_SITES,
+  DEFAULT_LIMIT_FOR_QUERYING_PHOTOS:
+    process.env.DEFAULT_LIMIT_FOR_QUERYING_PHOTOS,
+  DEFAULT_LIMIT_FOR_QUERYING_TIPS: process.env.DEFAULT_LIMIT_FOR_QUERYING_TIPS,
+  DEFAULT_LIMIT_FOR_QUERYING_KYA_TASKS:
+    process.env.DEFAULT_LIMIT_FOR_QUERYING_KYA_TASKS,
+  DEFAULT_LIMIT_FOR_QUERYING_KYA_LESSONS:
+    process.env.DEFAULT_LIMIT_FOR_QUERYING_KYA_LESSONS,
+  DEFAULT_LIMIT_FOR_QUERYING_KYA_QUESTIONS:
+    process.env.DEFAULT_LIMIT_FOR_QUERYING_KYA_QUESTIONS,
+  DEFAULT_LIMIT_FOR_QUERYING_KYA_ANSWERS:
+    process.env.DEFAULT_LIMIT_FOR_QUERYING_KYA_ANSWERS,
+  DEFAULT_LIMIT_FOR_QUERYING_KYA_QUIZZES:
+    process.env.DEFAULT_LIMIT_FOR_QUERYING_KYA_QUIZZES,
+  DEFAULT_LIMIT_FOR_QUERYING_AIRQLOUDS:
+    process.env.DEFAULT_LIMIT_FOR_QUERYING_AIRQLOUDS,
+  DEFAULT_EVENTS_LIMIT: process.env.DEFAULT_EVENTS_LIMIT,
+  DEFAULT_EVENTS_SKIP: process.env.DEFAULT_EVENTS_SKIP,
+  EVENTS_CACHE_LIMIT: process.env.EVENTS_CACHE_LIMIT,
+};
+module.exports = queryLimits;
